Fix antialias
